% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictxgb.R
\name{predict.xgb}
\alias{predict.xgb}
\title{Predict Method for XtremeBoost Objects}
\usage{
\method{predict}{xgb}(object, newdata, n.trees, ...)
}
\arguments{
\item{object}{an XtremeBoost object}

\item{newdata}{data required to perform}

\item{n.trees}{number of trees to use from the object.}

\item{...}{additional arguments affecting the predictions produced.}
}
\value{
a vector with predictions
}
\description{
It takes an XtremeBoost model from the \code{xboosting()} and performs a prediction
}
\references{
Marie Salditt, Sarah Humberg & Steffen Nestler (2023) Gradient Tree Boosting for Hierarchical Data, Multivariate Behavioral Research, 58:5, 911-937, DOI: 10.1080/00273171.2022.2146638
}
