% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xboosting.R
\name{boost_mem}
\alias{boost_mem}
\title{Iteration function for Tree-based MM}
\usage{
boost_mem(
  formula,
  data = NULL,
  random = NULL,
  shrinkage = 0.3,
  loss = "reg:squarederror",
  interaction.depth = 20,
  n.trees = 100,
  minsplit = 20,
  subsample = 0.5,
  conv_memboost = 0.001,
  maxIter_memboost = 100,
  minIter_memboost = 0,
  verbose_memboost = FALSE
)
}
\arguments{
\item{verbose_memboost}{}
}
\description{
Iteration function for Tree-based MM
}
