% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{mem_boost_gll}
\alias{mem_boost_gll}
\title{Log-Likelihood convergence function}
\usage{
mem_boost_gll(
  Z = NULL,
  ID = NULL,
  bhat = NULL,
  ehat = NULL,
  UniqueID = NULL,
  NID = NULL,
  D = NULL,
  Sigma2 = NULL
)
}
\arguments{
\item{Z}{An \eqn{n_i \times k} matrix representing the random effects}

\item{ID}{Domains of random effects}

\item{bhat}{random effects}

\item{ehat}{Error terms}

\item{UniqueID}{Unique identifier}

\item{NID}{Number of domains}

\item{D}{An \eqn{k \times k} covariance matrix}

\item{Sigma2}{Variance of error terms
\deqn{GLL(f, b_i|y_i) = \sum_{i=1}^n \left[ \left( y_i - f(X_i) - Z_i b_i \right)^T \Sigma_i^{-1} \left( y_i - f(X_i) - Z_i b_i \right) + b_i^T D^{-1} b_i + \log(\left| D \right|) + \log(\left| \Sigma_i \right|) \right]}}
}
\value{
Value of GLL
}
\description{
Function that takes the estimates of the \code{boost_mem()} and checks for convergence
}
\references{
Marie Salditt, Sarah Humberg & Steffen Nestler (2023) Gradient Tree Boosting for Hierarchical Data,
Multivariate Behavioral Research, 58:5, 911-937, DOI: 10.1080/00273171.2022.2146638

Krennmair, P., & Schmid, T. (2022). Flexible Domain Prediction using Mixed Effects Random Forests.
Applied Statistics/Journal Of The Royal Statistical Society. Series C, Applied Statistics, 71(5), 1865-1894. https://doi.org/10.1111/rssc.12600
}
