% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xboosting.R
\name{mem_boost_gll}
\alias{mem_boost_gll}
\title{GLL function}
\usage{
mem_boost_gll(
  Z = NULL,
  ID = NULL,
  bhat = NULL,
  ehat = NULL,
  UniqueID = NULL,
  NID = NULL,
  D = NULL,
  Sigma2 = NULL
)
}
\arguments{
\item{Sigma2}{}
}
\description{
GLL function
}
