% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xboosting.R
\name{mem_boost_gll}
\alias{mem_boost_gll}
\title{Log-Likelihood convergence function}
\usage{
mem_boost_gll(
  Z = NULL,
  ID = NULL,
  bhat = NULL,
  ehat = NULL,
  UniqueID = NULL,
  NID = NULL,
  D = NULL,
  Sigma2 = NULL
)
}
\arguments{
\item{Sigma2}{}
}
\description{
Function that takes the estimates of the `boost_mem()` and checks for convergence
}
\references{
Marie Salditt, Sarah Humberg & Steffen Nestler (2023) Gradient Tree Boosting for Hierarchical Data, Multivariate Behavioral Research, 58:5, 911-937, DOI: 10.1080/00273171.2022.2146638
}
