% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation}
\alias{validation}
\title{Direct Estimate Validation}
\usage{
validation(
  test = NULL,
  prediction = NULL,
  validation_set = NULL,
  weights = NULL,
  label = NULL,
  region = NULL,
  model = NULL
)
}
\arguments{
\item{test}{Test set used in the prediction of the fitted values (usually administrative records)}

\item{prediction}{Vector type object with the fitted values}

\item{validation_set}{Data set to fit Direct Estimates}

\item{weights}{Name of column with the weights for direct estimates}

\item{label}{response variable}

\item{region}{domain variable}

\item{model}{Model name for the plot}
}
\value{
Returns a list of values and a \code{ggplot()} object with the Direct Estimates.
}
\description{
A function that fits the Direct Estimates in Small Area Estimation
and plots it against the estimates of the model.
}
