% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xboosting.R
\name{xboosting}
\alias{xboosting}
\title{Xboosting}
\usage{
xboosting(
  formula,
  data = NULL,
  loss = "reg:squarederror",
  n.trees = 100,
  shrinkage = 0.1,
  interaction.depth = 1,
  minsplit = 20,
  subsample = 0.5
)
}
\arguments{
\item{subsample}{}
}
\description{
Xboosting
}
